% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_pmt}
\alias{read_pmt}
\title{A function that imports all payment}
\usage{
read_pmt(prod_month)
}
\arguments{
\item{prod_month}{The production month}
}
\value{
A data frame of all payment data
}
\description{
A function that imports all payment
}
\examples{
prod_month <- ymd('2024-11-01')
read_pmt(prod_month)
}
